(fix|COS-4003|COS-4046): Fixed sli name saving issues + removed redundant calls after closing item + decrease invalidate calls and increase timeout for the ones that are left out
