--- packages/apps/frontera/src/routes/organization/src/components/Tabs/panels/AccountPanel/Contract/ContractBillingDetailsModal/Services/components/highlighters/index.ts
@@ -1,2 +1,0 @@
-export { Highlighter } from './Highlighter';
-export { getVersionFromUUID, getColorByUUID } from './utils';
