--- packages/apps/frontera/src/routes/organization/src/components/Tabs/panels/AccountPanel/Contract/ContractBillingDetailsModal/Services/index.ts
@@ -1,1 +1,0 @@
-export { Services } from './Services';
