+++ packages/apps/frontera/src/routes/organization/src/components/Tabs/panels/AccountPanel/Contract/ContractBillingDetailsModal/components/BillingAddressDetails/index.ts
@@ -1,0 +1,1 @@
+export { BillingAddressModal } from './BillingAddressModal.tsx';
