+++ packages/apps/frontera/src/routes/organization/src/components/Tabs/panels/AccountPanel/Contract/ContractBillingDetailsModal/components/Services/components/ServiceList.tsx
@@ -7,7 +7,7 @@ import { ContractLineItemStore } from '@store/ContractLineItems/ContractLineItem
 import { useStore } from '@shared/hooks/useStore';
 import { BilledType, ContractStatus, ServiceLineItem } from '@graphql/types';
 
-import { ServiceCard } from './ServiceCard';
+import { ServiceCard } from './ServiceCard.tsx';
 
 export const ServiceList: React.FC<{
   id: string;
