+++ packages/apps/frontera/src/routes/organization/src/components/Tabs/panels/AccountPanel/Contract/ContractBillingDetailsModal/components/Services/components/highlighters/index.ts
@@ -1,0 +1,2 @@
+export { Highlighter } from './Highlighter.tsx';
+export { getVersionFromUUID, getColorByUUID } from './utils.ts';
