+++ packages/apps/frontera/src/routes/organization/src/components/Tabs/panels/AccountPanel/Contract/ContractBillingDetailsModal/components/Services/index.ts
@@ -1,0 +1,1 @@
+export { Services } from './Services.tsx';
