+++ packages/apps/frontera/src/store/ContractLineItems/ContractLineItem.store.ts
@@ -110,7 +110,6 @@ export class ContractLineItemStore implements Store<ServiceLineItem> {
     } finally {
       runInAction(() => {
         this.isLoading = false;
-        this.invalidate();
       });
     }
   }
